fix(clojure): Fix visibility for compatibility with Clojure
Closes #451

This change put all internal interfaces public so it can be access via reflexivity.
It's a work around for this bug: https://bugs.java.com/bugdatabase/view_bug.do?bug_id=4283544
