+++ algoliasearch-common/src/main/java/com/algolia/search/AsyncBaseIndex.java
@@ -1,0 +1,6 @@
+package com.algolia.search;
+
+public interface AsyncBaseIndex<T> extends AbstractIndex<T> {
+
+  AsyncAPIClient getApiClient();
+}
