+++ algoliasearch-common/src/main/java/com/algolia/search/SyncBaseIndex.java
@@ -1,0 +1,6 @@
+package com.algolia.search;
+
+public interface SyncBaseIndex<T> extends AbstractIndex<T> {
+
+  APIClient getApiClient();
+}
